refactor: rename activeCellForTableText to cellAnchors
The module no longer derives anything from table text: its exports are
computeCellAnchorFromRanges and computeCellAnchorForTable, both of which
take an already-derived model or ranges.
